Fix up and simplify
